apps/requirements.txt
# Core dependencies
numpy>=1.21.0
opencv-python>=4.5.0
pillow>=8.0.0
ultralytics>=8.0.0
onnxruntime>=1.12.0
mss>=6.1.0
pynput>=1.7.0
pyyaml>=6.0

# macOS specific (optional)
pyobjc-framework-Quartz>=8.0; sys_platform == "darwin"

# Development dependencies (optional)
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0
